+++ src/Accounts/Accounts/StartupScripts/InitializeAssemblyResolver.ps1
@@ -8,6 +8,7 @@
 }
 else {
   try {
+    Add-Type -Path ([System.IO.Path]::Combine($PSScriptRoot, "..", "Microsoft.Azure.PowerShell.AuthenticationAssemblyLoadContext.dll")) | Out-Null
     $assemblyLoadContextFolder = [System.IO.Path]::Combine($PSScriptRoot, "..", "AzSharedAlcAssemblies")
     Write-Debug "Registering Az shared AssemblyLoadContext for path: '$assemblyLoadContextFolder'."
     [Microsoft.Azure.PowerShell.AuthenticationAssemblyLoadContext.AzAssemblyLoadContextInitializer]::RegisterAzSharedAssemblyLoadContext($assemblyLoadContextFolder)