+++ src/Resources/MSGraph.Autorest/internal/Get-AzADUser.ps1
@@ -156,8 +156,8 @@ begin {
         }
         $parameterSet = $PSCmdlet.ParameterSetName
         $mapping = @{
-            Get = 'Az.Resources.MSGraph.private\Get-AzADUser_Get';
-            List = 'Az.Resources.MSGraph.private\Get-AzADUser_List';
+            Get = 'Az.MSGraph.private\Get-AzADUser_Get';
+            List = 'Az.MSGraph.private\Get-AzADUser_List';
         }
 
         $wrappedCmd = $ExecutionContext.InvokeCommand.GetCommand(($mapping[$parameterSet]), [System.Management.Automation.CommandTypes]::Cmdlet)