+++ src/Resources/MSGraph.Autorest/internal/Remove-AzADServicePrincipalPassword.ps1
@@ -134,8 +134,8 @@ begin {
         }
         $parameterSet = $PSCmdlet.ParameterSetName
         $mapping = @{
-            Remove = 'Az.Resources.MSGraph.private\Remove-AzADServicePrincipalPassword_Remove';
-            RemoveExpanded = 'Az.Resources.MSGraph.private\Remove-AzADServicePrincipalPassword_RemoveExpanded';
+            Remove = 'Az.MSGraph.private\Remove-AzADServicePrincipalPassword_Remove';
+            RemoveExpanded = 'Az.MSGraph.private\Remove-AzADServicePrincipalPassword_RemoveExpanded';
         }
 
         $wrappedCmd = $ExecutionContext.InvokeCommand.GetCommand(($mapping[$parameterSet]), [System.Management.Automation.CommandTypes]::Cmdlet)