+++ src/Storage/Storage/Common/Cmdlet/GetAzureStorageServiceProperties.cs
@@ -20,6 +20,7 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Common.Cmdlet
     using System.Security.Permissions;
     using Microsoft.WindowsAzure.Commands.Storage.Model.ResourceModel;
     using Microsoft.WindowsAzure.Commands.Storage.Model.Contract;
+    using global::Azure.Data.Tables.Models;
 
     /// <summary>
     /// Show Azure Storage service properties
@@ -55,8 +56,17 @@ namespace Microsoft.WindowsAzure.Commands.Storage.Common.Cmdlet
             else //Table use old XSCL
             {
                 StorageTableManagement tableChannel = new StorageTableManagement(Channel.StorageContext);
-                XTable.ServiceProperties serviceProperties = tableChannel.GetStorageTableServiceProperties(GetTableRequestOptions(), TableOperationContext);
-                WriteObject(new PSSeriviceProperties(serviceProperties));
+
+                if (!tableChannel.IsTokenCredential)
+                {
+                    XTable.ServiceProperties serviceProperties = tableChannel.GetStorageTableServiceProperties(GetTableRequestOptions(), TableOperationContext);
+                    WriteObject(new PSSeriviceProperties(serviceProperties));
+                }
+                else
+                {
+                    TableServiceProperties serviceProperties = tableChannel.GetProperties(this.CmdletCancellationToken);
+                    WriteObject(new PSSeriviceProperties(serviceProperties));
+                }
             }
         }
     }