+++ src/Compute/Compute/ChangeLog.md
@@ -20,9 +20,10 @@
 
 -->
 ## Upcoming Release
-* Changed `New-AzVM` cmdlet when using the SimpleParameterSet to not create a PublicIPAddress when a PublicIPAddress name is not provided. 
-* Add `PlatformFaultDomain` parameter to cmdlets: `New-AzVM` and `New-AzVMConfig`
+* Changed `New-AzVM` cmdlet when using the SimpleParameterSet to not create a `PublicIPAddress` when a `PublicIPAddress` name is not provided.
+* Added `PlatformFaultDomain` parameter to cmdlets: `New-AzVM` and `New-AzVMConfig`
 * Added `-Feature` parameter for `New-AzGalleryImageDefinition`
+* Added `DiffDiskPlacement` string parameter to `Set-AzVmOSDisk` and `Set-AzVmssStorageProfile` cmdlets.
 * Updated `UserData` parameter in VM and VMSS cmdlets to pipe by the Property Name to ensure piping scenarios occur correctly.`
 
 ## Version 4.21.0